refactor: reparando caso de duplicação de código e separando as responsabilidades dos métodos cadastrarProduto e editarProduto
